model fix

+++ lms/course/models.py
@@ -24,8 +24,8 @@ class Tasks(models.Model):
     close_assert = models.TextField(blank=True, null=True, verbose_name='Закрытые тесты')
     slug = models.SlugField(max_length=200, unique=True, blank=True, null=True, verbose_name='URL')
     description = models.TextField(blank=True, null=True, verbose_name='Описание')
-    time = models.DecimalField(default=0.00, decimal_places=2, max_digits=7, verbose_name='Время')
-    memory = models.DecimalField(default=0.00, decimal_places=2, max_digits=7, verbose_name='Память')
+    time = models.FloatField(default=0.00, verbose_name='Время')
+    memory = models.FloatField(default=0.00, verbose_name='Память')
     # memory_unit = models.CharField(max_length=15, verbose_name='Единицы измерения памяти')
     course = models.ForeignKey(to=Courses, on_delete=models.CASCADE)
